+++ prov/efa/Makefile.include
@@ -61,6 +61,7 @@ _efa_files = \
 	prov/efa/src/rdm/efa_rdm_pke.c \
 	prov/efa/src/rdm/efa_rdm_pke_utils.c \
 	prov/efa/src/rdm/efa_rdm_pke_rtr.c \
+	prov/efa/src/rdm/efa_rdm_pke_rtw.c \
 	prov/efa/src/rdm/efa_rdm_pke_rta.c \
 	prov/efa/src/rdm/rxr_pkt_type_req.c \
 	prov/efa/src/rdm/efa_rdm_pke_nonreq.c \
@@ -102,6 +103,7 @@ _efa_headers = \
 	prov/efa/src/rdm/efa_rdm_pkt_type.h \
 	prov/efa/src/rdm/rxr_pkt_type_req.h \
 	prov/efa/src/rdm/efa_rdm_pke_rtr.h \
+	prov/efa/src/rdm/efa_rdm_pke_rtw.h \
 	prov/efa/src/rdm/efa_rdm_pke_rta.h \
 	prov/efa/src/rdm/efa_rdm_pke_nonreq.h \
 	prov/efa/src/rdm/efa_rdm_pke_cmd.h \
