+++ prov/efa/src/rdm/rxr_pkt_type_req.c
@@ -41,6 +41,7 @@
 #include "efa_rdm_srx.h"
 #include "rxr_pkt_type_req.h"
 #include "efa_rdm_pke_rta.h"
+#include "efa_rdm_pke_rtw.h"
 #include "efa_rdm_pke_rtr.h"
 #include "rxr_tp.h"
 
@@ -218,7 +219,7 @@ uint32_t rxr_pkt_hdr_rma_iov_count(struct efa_rdm_pke *pkt_entry)
 	    pkt_type == RXR_LONGCTS_RTW_PKT ||
 	    pkt_type == RXR_DC_LONGCTS_RTW_PKT ||
 	    pkt_type == RXR_LONGREAD_RTW_PKT)
-		return rxr_get_rtw_base_hdr(pkt_entry->wiredata)->rma_iov_count;
+		return efa_rdm_pke_get_rtw_base_hdr(pkt_entry)->rma_iov_count;
 
 	if (pkt_type == RXR_SHORT_RTR_PKT ||
 		 pkt_type == RXR_LONGCTS_RTR_PKT)
@@ -1561,420 +1562,4 @@ void rxr_pkt_handle_rtm_rta_recv(struct efa_rdm_ep *ep,
 
 	ofi_recvwin_slide((&peer->robuf));
 	efa_rdm_peer_proc_pending_items_in_robuf(peer, ep);
-}
-
-/*
- * RTW packet type functions
- */
-static inline
-ssize_t rxr_pkt_init_rtw_data(struct efa_rdm_pke *pkt_entry,
-			      struct efa_rdm_ope *txe,
-			      struct efa_rma_iov *rma_iov)
-{
-	size_t hdr_size;
-	size_t data_size;
-	int i;
-
-	for (i = 0; i < txe->rma_iov_count; ++i) {
-		rma_iov[i].addr = txe->rma_iov[i].addr;
-		rma_iov[i].len = txe->rma_iov[i].len;
-		rma_iov[i].key = txe->rma_iov[i].key;
-	}
-
-	hdr_size = rxr_pkt_req_hdr_size_from_pkt_entry(pkt_entry);
-	data_size = MIN(txe->ep->mtu_size - hdr_size, txe->total_len);
-	return efa_rdm_pke_init_payload_from_ope(pkt_entry, txe, hdr_size, 0, data_size);
-}
-
-ssize_t rxr_pkt_init_eager_rtw(struct efa_rdm_pke *pkt_entry,
-			       struct efa_rdm_ope *txe)
-{
-	struct rxr_eager_rtw_hdr *rtw_hdr;
-
-	assert(txe->op == ofi_op_write);
-
-	rtw_hdr = (struct rxr_eager_rtw_hdr *)pkt_entry->wiredata;
-	rtw_hdr->rma_iov_count = txe->rma_iov_count;
-	rxr_pkt_init_req_hdr(pkt_entry, RXR_EAGER_RTW_PKT, txe);
-	return rxr_pkt_init_rtw_data(pkt_entry, txe, rtw_hdr->rma_iov);
-}
-
-ssize_t rxr_pkt_init_dc_eager_rtw(struct efa_rdm_pke *pkt_entry,
-				  struct efa_rdm_ope *txe)
-{
-	struct rxr_dc_eager_rtw_hdr *dc_eager_rtw_hdr;
-	int ret;
-
-	assert(txe->op == ofi_op_write);
-
-	txe->rxr_flags |= EFA_RDM_TXE_DELIVERY_COMPLETE_REQUESTED;
-	dc_eager_rtw_hdr = (struct rxr_dc_eager_rtw_hdr *)pkt_entry->wiredata;
-	dc_eager_rtw_hdr->rma_iov_count = txe->rma_iov_count;
-	rxr_pkt_init_req_hdr(pkt_entry, RXR_DC_EAGER_RTW_PKT, txe);
-	ret = rxr_pkt_init_rtw_data(pkt_entry, txe,
-				    dc_eager_rtw_hdr->rma_iov);
-	dc_eager_rtw_hdr->send_id = txe->tx_id;
-	return ret;
-}
-
-static inline
-void rxr_pkt_init_longcts_rtw_hdr(struct efa_rdm_pke *pkt_entry,
-				  int pkt_type,
-				  struct efa_rdm_ope *txe)
-{
-	struct rxr_longcts_rtw_hdr *rtw_hdr;
-
-	rtw_hdr = (struct rxr_longcts_rtw_hdr *)pkt_entry->wiredata;
-	rtw_hdr->rma_iov_count = txe->rma_iov_count;
-	rtw_hdr->msg_length = txe->total_len;
-	rtw_hdr->send_id = txe->tx_id;
-	rtw_hdr->credit_request = efa_env.tx_min_credits;
-	rxr_pkt_init_req_hdr(pkt_entry, pkt_type, txe);
-}
-
-ssize_t rxr_pkt_init_longcts_rtw(struct efa_rdm_pke *pkt_entry,
-				 struct efa_rdm_ope *txe)
-{
-	struct rxr_longcts_rtw_hdr *rtw_hdr;
-
-	assert(txe->op == ofi_op_write);
-
-	rtw_hdr = (struct rxr_longcts_rtw_hdr *)pkt_entry->wiredata;
-	rxr_pkt_init_longcts_rtw_hdr(pkt_entry, RXR_LONGCTS_RTW_PKT, txe);
-	return rxr_pkt_init_rtw_data(pkt_entry, txe, rtw_hdr->rma_iov);
-}
-
-ssize_t rxr_pkt_init_dc_longcts_rtw(struct efa_rdm_pke *pkt_entry,
-				    struct efa_rdm_ope *txe)
-{
-	struct rxr_longcts_rtw_hdr *rtw_hdr;
-
-	assert(txe->op == ofi_op_write);
-
-	txe->rxr_flags |= EFA_RDM_TXE_DELIVERY_COMPLETE_REQUESTED;
-	rtw_hdr = (struct rxr_longcts_rtw_hdr *)pkt_entry->wiredata;
-	rxr_pkt_init_longcts_rtw_hdr(pkt_entry, RXR_DC_LONGCTS_RTW_PKT, txe);
-	return rxr_pkt_init_rtw_data(pkt_entry, txe, rtw_hdr->rma_iov);
-}
-
-ssize_t rxr_pkt_init_longread_rtw(struct efa_rdm_pke *pkt_entry,
-				  struct efa_rdm_ope *txe)
-{
-	struct rxr_longread_rtw_hdr *rtw_hdr;
-	struct efa_rma_iov *rma_iov;
-	struct fi_rma_iov *read_iov;
-	size_t hdr_size;
-	int i, err;
-
-	assert(txe->op == ofi_op_write);
-
-	rtw_hdr = (struct rxr_longread_rtw_hdr *)pkt_entry->wiredata;
-	rtw_hdr->rma_iov_count = txe->rma_iov_count;
-	rtw_hdr->msg_length = txe->total_len;
-	rtw_hdr->send_id = txe->tx_id;
-	rtw_hdr->read_iov_count = txe->iov_count;
-	rxr_pkt_init_req_hdr(pkt_entry, RXR_LONGREAD_RTW_PKT, txe);
-
-	rma_iov = rtw_hdr->rma_iov;
-	for (i = 0; i < txe->rma_iov_count; ++i) {
-		rma_iov[i].addr = txe->rma_iov[i].addr;
-		rma_iov[i].len = txe->rma_iov[i].len;
-		rma_iov[i].key = txe->rma_iov[i].key;
-	}
-
-	hdr_size = rxr_pkt_req_hdr_size_from_pkt_entry(pkt_entry);
-	read_iov = (struct fi_rma_iov *)(pkt_entry->wiredata + hdr_size);
-	err = efa_rdm_txe_prepare_to_be_read(txe, read_iov);
-	if (OFI_UNLIKELY(err))
-		return err;
-
-	pkt_entry->pkt_size = hdr_size + txe->iov_count * sizeof(struct efa_rma_iov);
-	pkt_entry->ope = txe;
-	return 0;
-}
-
-/*
- *     handle_sent() functions for RTW packet types
- *
- *         rxr_pkt_handle_longcts_rtw_sent() is empty and is defined in rxr_pkt_type_req.h
- */
-void rxr_pkt_handle_longcts_rtw_sent(struct efa_rdm_ep *ep,
-				  struct efa_rdm_pke *pkt_entry)
-{
-	struct efa_rdm_ope *txe;
-	struct efa_domain *efa_domain = efa_rdm_ep_domain(ep);
-
-	txe = pkt_entry->ope;
-	txe->bytes_sent += pkt_entry->payload_size;
-	assert(txe->bytes_sent < txe->total_len);
-	if (efa_is_cache_available(efa_domain))
-		efa_rdm_ope_try_fill_desc(txe, 0, FI_SEND);
-}
-
-/*
- *     handle_send_completion() functions
- */
-void rxr_pkt_handle_eager_rtw_send_completion(struct efa_rdm_ep *ep,
-					      struct efa_rdm_pke *pkt_entry)
-{
-	struct efa_rdm_ope *txe;
-
-	txe = pkt_entry->ope;
-	assert(txe->total_len == pkt_entry->payload_size);
-	efa_rdm_ope_handle_send_completed(txe);
-}
-
-void rxr_pkt_handle_longcts_rtw_send_completion(struct efa_rdm_ep *ep,
-					     struct efa_rdm_pke *pkt_entry)
-{
-	struct efa_rdm_ope *txe;
-
-	txe = pkt_entry->ope;
-	txe->bytes_acked += pkt_entry->payload_size;
-	if (txe->total_len == txe->bytes_acked)
-		efa_rdm_ope_handle_send_completed(txe);
-}
-
-/*
- *     handle_recv() functions
- */
-
-static
-struct efa_rdm_ope *rxr_pkt_alloc_rtw_rxe(struct efa_rdm_ep *ep,
-						struct efa_rdm_pke *pkt_entry)
-{
-	struct efa_rdm_ope *rxe;
-	struct rxr_base_hdr *base_hdr;
-
-	rxe = efa_rdm_ep_alloc_rxe(ep, pkt_entry->addr, ofi_op_write);
-	if (OFI_UNLIKELY(!rxe))
-		return NULL;
-
-	base_hdr = efa_rdm_pke_get_base_hdr(pkt_entry);
-	if (base_hdr->flags & RXR_REQ_OPT_CQ_DATA_HDR) {
-		rxe->cq_entry.flags |= FI_REMOTE_CQ_DATA;
-		rxe->cq_entry.data = rxr_pkt_req_cq_data(pkt_entry);
-	}
-
-	rxe->addr = pkt_entry->addr;
-	rxe->bytes_received = 0;
-	rxe->bytes_copied = 0;
-	return rxe;
-}
-
-void rxr_pkt_proc_eager_rtw(struct efa_rdm_ep *ep,
-			    struct efa_rma_iov *rma_iov,
-			    size_t rma_iov_count,
-			    struct efa_rdm_ope *rxe,
-			    struct efa_rdm_pke *pkt_entry)
-{
-	ssize_t err;
-
-	err = efa_rdm_rma_verified_copy_iov(ep, rma_iov, rma_iov_count,
-					FI_REMOTE_WRITE, rxe->iov, rxe->desc);
-
-	if (OFI_UNLIKELY(err)) {
-		EFA_WARN(FI_LOG_CQ, "RMA address verify failed!\n");
-		efa_base_ep_write_eq_error(&ep->base_ep, FI_EIO, FI_EFA_ERR_RMA_ADDR);
-		efa_rdm_rxe_release(rxe);
-		efa_rdm_pke_release_rx(ep, pkt_entry);
-		return;
-	}
-
-	rxe->cq_entry.len = ofi_total_iov_len(rxe->iov, rxe->iov_count);
-	rxe->cq_entry.buf = rxe->iov[0].iov_base;
-	rxe->total_len = rxe->cq_entry.len;
-
-	rxe->bytes_received += pkt_entry->payload_size;
-	if (pkt_entry->payload_size != rxe->total_len) {
-		EFA_WARN(FI_LOG_CQ, "Eager RTM size mismatch! payload_size: %ld total_len: %ld.\n",
-			 pkt_entry->payload_size, rxe->total_len);
-		EFA_WARN(FI_LOG_CQ, "target buffer: %p length: %ld\n", rxe->iov[0].iov_base,
-			rxe->iov[0].iov_len);
-		efa_base_ep_write_eq_error(&ep->base_ep, FI_EINVAL, FI_EFA_ERR_RTM_MISMATCH);
-		efa_rdm_pke_release_rx(ep, pkt_entry);
-		efa_rdm_rxe_release(rxe);
-	} else {
-		err = efa_rdm_pke_copy_payload_to_ope(pkt_entry, rxe);
-		if (OFI_UNLIKELY(err)) {
-			efa_base_ep_write_eq_error(&ep->base_ep, FI_EINVAL, FI_EFA_ERR_RXE_COPY);
-			efa_rdm_pke_release_rx(ep, pkt_entry);
-			efa_rdm_rxe_release(rxe);
-		}
-	}
-}
-
-void rxr_pkt_handle_eager_rtw_recv(struct efa_rdm_ep *ep,
-				   struct efa_rdm_pke *pkt_entry)
-{
-	struct efa_rdm_ope *rxe;
-	struct rxr_eager_rtw_hdr *rtw_hdr;
-
-	rxe = rxr_pkt_alloc_rtw_rxe(ep, pkt_entry);
-
-	if (!rxe) {
-		EFA_WARN(FI_LOG_CQ,
-			"RX entries exhausted.\n");
-		efa_base_ep_write_eq_error(&ep->base_ep, FI_ENOBUFS, FI_EFA_ERR_RXE_POOL_EXHAUSTED);
-		efa_rdm_pke_release_rx(ep, pkt_entry);
-		return;
-	}
-
-	rtw_hdr = (struct rxr_eager_rtw_hdr *)pkt_entry->wiredata;
-	rxe->iov_count = rtw_hdr->rma_iov_count;
-	rxr_pkt_proc_eager_rtw(ep,
-			       rtw_hdr->rma_iov,
-			       rtw_hdr->rma_iov_count,
-			       rxe, pkt_entry);
-}
-
-void rxr_pkt_handle_dc_eager_rtw_recv(struct efa_rdm_ep *ep,
-				      struct efa_rdm_pke *pkt_entry)
-{
-	struct efa_rdm_ope *rxe;
-	struct rxr_dc_eager_rtw_hdr *rtw_hdr;
-
-	rxe = rxr_pkt_alloc_rtw_rxe(ep, pkt_entry);
-	if (!rxe) {
-		EFA_WARN(FI_LOG_CQ,
-			"RX entries exhausted.\n");
-		efa_base_ep_write_eq_error(&ep->base_ep, FI_ENOBUFS, FI_EFA_ERR_RXE_POOL_EXHAUSTED);
-		efa_rdm_pke_release_rx(ep, pkt_entry);
-		return;
-	}
-
-	rxe->rxr_flags |= EFA_RDM_TXE_DELIVERY_COMPLETE_REQUESTED;
-	rtw_hdr = (struct rxr_dc_eager_rtw_hdr *)pkt_entry->wiredata;
-	rxe->tx_id = rtw_hdr->send_id;
-	rxe->iov_count = rtw_hdr->rma_iov_count;
-	rxr_pkt_proc_eager_rtw(ep,
-			       rtw_hdr->rma_iov,
-			       rtw_hdr->rma_iov_count,
-			       rxe, pkt_entry);
-}
-
-void rxr_pkt_handle_longcts_rtw_recv(struct efa_rdm_ep *ep,
-				  struct efa_rdm_pke *pkt_entry)
-{
-	struct efa_rdm_ope *rxe;
-	struct rxr_longcts_rtw_hdr *rtw_hdr;
-	ssize_t err;
-	uint32_t tx_id;
-
-	rxe = rxr_pkt_alloc_rtw_rxe(ep, pkt_entry);
-	if (!rxe) {
-		EFA_WARN(FI_LOG_CQ,
-			"RX entries exhausted.\n");
-		efa_base_ep_write_eq_error(&ep->base_ep, FI_ENOBUFS, FI_EFA_ERR_RXE_POOL_EXHAUSTED);
-		efa_rdm_pke_release_rx(ep, pkt_entry);
-		return;
-	}
-
-	rtw_hdr = (struct rxr_longcts_rtw_hdr *)pkt_entry->wiredata;
-	tx_id = rtw_hdr->send_id;
-	if (rtw_hdr->type == RXR_DC_LONGCTS_RTW_PKT)
-		rxe->rxr_flags |= EFA_RDM_TXE_DELIVERY_COMPLETE_REQUESTED;
-
-	rxe->iov_count = rtw_hdr->rma_iov_count;
-	err = efa_rdm_rma_verified_copy_iov(ep, rtw_hdr->rma_iov, rtw_hdr->rma_iov_count,
-					FI_REMOTE_WRITE, rxe->iov, rxe->desc);
-	if (OFI_UNLIKELY(err)) {
-		EFA_WARN(FI_LOG_CQ, "RMA address verify failed!\n");
-		efa_base_ep_write_eq_error(&ep->base_ep, FI_EIO, FI_EFA_ERR_RMA_ADDR);
-		efa_rdm_rxe_release(rxe);
-		efa_rdm_pke_release_rx(ep, pkt_entry);
-		return;
-	}
-
-	rxe->cq_entry.len = ofi_total_iov_len(rxe->iov, rxe->iov_count);
-	rxe->cq_entry.buf = rxe->iov[0].iov_base;
-	rxe->total_len = rxe->cq_entry.len;
-
-	rxe->bytes_received += pkt_entry->payload_size;
-	if (pkt_entry->payload_size >= rxe->total_len) {
-		EFA_WARN(FI_LOG_CQ, "Long RTM size mismatch! payload_size: %ld total_len: %ld\n",
-			 pkt_entry->payload_size, rxe->total_len);
-		EFA_WARN(FI_LOG_CQ, "target buffer: %p length: %ld\n", rxe->iov[0].iov_base,
-			rxe->iov[0].iov_len);
-		efa_base_ep_write_eq_error(&ep->base_ep, FI_EINVAL, FI_EFA_ERR_RTM_MISMATCH);
-		efa_rdm_rxe_release(rxe);
-		efa_rdm_pke_release_rx(ep, pkt_entry);
-		return;
-	} else {
-		err = efa_rdm_pke_copy_payload_to_ope(pkt_entry, rxe);
-		if (OFI_UNLIKELY(err)) {
-			efa_base_ep_write_eq_error(&ep->base_ep, FI_EINVAL, FI_EFA_ERR_RXE_COPY);
-			efa_rdm_rxe_release(rxe);
-			efa_rdm_pke_release_rx(ep, pkt_entry);
-			return;
-		}
-	}
-
-
-#if ENABLE_DEBUG
-	dlist_insert_tail(&rxe->pending_recv_entry, &ep->ope_recv_list);
-	ep->pending_recv_counter++;
-#endif
-	rxe->state = EFA_RDM_RXE_RECV;
-	rxe->tx_id = tx_id;
-	err = efa_rdm_ope_post_send_or_queue(rxe, RXR_CTS_PKT);
-	if (OFI_UNLIKELY(err)) {
-		EFA_WARN(FI_LOG_CQ, "Cannot post CTS packet\n");
-		efa_rdm_rxe_handle_error(rxe, -err, FI_EFA_ERR_PKT_POST);
-		efa_rdm_rxe_release(rxe);
-	}
-}
-
-void rxr_pkt_handle_longread_rtw_recv(struct efa_rdm_ep *ep,
-				  struct efa_rdm_pke *pkt_entry)
-{
-	struct efa_rdm_ope *rxe;
-	struct rxr_longread_rtw_hdr *rtw_hdr;
-	struct fi_rma_iov *read_iov;
-	size_t hdr_size;
-	ssize_t err;
-
-	rxe = rxr_pkt_alloc_rtw_rxe(ep, pkt_entry);
-	if (!rxe) {
-		EFA_WARN(FI_LOG_CQ,
-			"RX entries exhausted.\n");
-		efa_base_ep_write_eq_error(&ep->base_ep, FI_ENOBUFS, FI_EFA_ERR_RXE_POOL_EXHAUSTED);
-		efa_rdm_pke_release_rx(ep, pkt_entry);
-		return;
-	}
-
-	rtw_hdr = (struct rxr_longread_rtw_hdr *)pkt_entry->wiredata;
-	rxe->iov_count = rtw_hdr->rma_iov_count;
-	err = efa_rdm_rma_verified_copy_iov(ep, rtw_hdr->rma_iov, rtw_hdr->rma_iov_count,
-					FI_REMOTE_WRITE, rxe->iov, rxe->desc);
-	if (OFI_UNLIKELY(err)) {
-		EFA_WARN(FI_LOG_CQ, "RMA address verify failed!\n");
-		efa_base_ep_write_eq_error(&ep->base_ep, FI_EINVAL, FI_EFA_ERR_RMA_ADDR);
-		efa_rdm_rxe_release(rxe);
-		efa_rdm_pke_release_rx(ep, pkt_entry);
-		return;
-	}
-
-	rxe->cq_entry.len = ofi_total_iov_len(rxe->iov, rxe->iov_count);
-	rxe->cq_entry.buf = rxe->iov[0].iov_base;
-	rxe->total_len = rxe->cq_entry.len;
-
-	hdr_size = rxr_pkt_req_hdr_size_from_pkt_entry(pkt_entry);
-	read_iov = (struct fi_rma_iov *)(pkt_entry->wiredata + hdr_size);
-	rxe->addr = pkt_entry->addr;
-	rxe->tx_id = rtw_hdr->send_id;
-	rxe->rma_iov_count = rtw_hdr->read_iov_count;
-	memcpy(rxe->rma_iov, read_iov,
-	       rxe->rma_iov_count * sizeof(struct fi_rma_iov));
-
-	efa_rdm_pke_release_rx(ep, pkt_entry);
-
-	err = efa_rdm_ope_post_remote_read_or_queue(rxe);
-	if (OFI_UNLIKELY(err)) {
-		EFA_WARN(FI_LOG_CQ,
-			"RDMA post read or queue failed.\n");
-		efa_base_ep_write_eq_error(&ep->base_ep, err, FI_EFA_ERR_RDMA_READ_POST);
-		efa_rdm_rxe_release(rxe);
-		efa_rdm_pke_release_rx(ep, pkt_entry);
-	}
 }
