+++ prov/efa/src/rdm/rxr_pkt_type_req.h
@@ -291,74 +291,6 @@ void rxr_pkt_handle_zcpy_recv(struct efa_rdm_ep *ep,
  */
 void rxr_pkt_handle_rtm_rta_recv(struct efa_rdm_ep *ep,
 				 struct efa_rdm_pke *pkt_entry);
-
-static inline
-struct rxr_rtw_base_hdr *rxr_get_rtw_base_hdr(void *pkt)
-{
-	return (struct rxr_rtw_base_hdr *)pkt;
-}
-
-static inline
-struct rxr_dc_eager_rtw_hdr *rxr_get_dc_eager_rtw_hdr(void *pkt)
-{
-	return (struct rxr_dc_eager_rtw_hdr *)pkt;
-}
-
-ssize_t rxr_pkt_init_eager_rtw(struct efa_rdm_pke *pkt_entry,
-			       struct efa_rdm_ope *txe);
-
-ssize_t rxr_pkt_init_longcts_rtw(struct efa_rdm_pke *pkt_entry,
-				 struct efa_rdm_ope *txe);
-
-ssize_t rxr_pkt_init_longread_rtw(struct efa_rdm_pke *pkt_entry,
-				  struct efa_rdm_ope *txe);
-
-ssize_t rxr_pkt_init_dc_eager_rtw(struct efa_rdm_pke *pkt_entry,
-				  struct efa_rdm_ope *txe);
-
-ssize_t rxr_pkt_init_dc_longcts_rtw(struct efa_rdm_pke *pkt_entry,
-				    struct efa_rdm_ope *txe);
-
-static inline
-void rxr_pkt_handle_eager_rtw_sent(struct efa_rdm_ep *ep,
-				   struct efa_rdm_pke *pkt_entry)
-{
-	/* For eager RTW, there is nothing to be done here */
-	return;
-}
-
-void rxr_pkt_handle_longcts_rtw_sent(struct efa_rdm_ep *ep,
-				  struct efa_rdm_pke *pkt_entry);
-
-static inline
-void rxr_pkt_handle_longread_rtw_sent(struct efa_rdm_ep *ep,
-				  struct efa_rdm_pke *pkt_entry)
-{
-}
-
-void rxr_pkt_handle_eager_rtw_send_completion(struct efa_rdm_ep *ep,
-					      struct efa_rdm_pke *pkt_entry);
-
-void rxr_pkt_handle_longcts_rtw_send_completion(struct efa_rdm_ep *ep,
-					     struct efa_rdm_pke *pkt_entry);
-
-static inline
-void rxr_pkt_handle_longread_rtw_send_completion(struct efa_rdm_ep *ep,
-					     struct efa_rdm_pke *pkt_entry)
-{
-}
-
-void rxr_pkt_handle_eager_rtw_recv(struct efa_rdm_ep *ep,
-				   struct efa_rdm_pke *pkt_entry);
-
-void rxr_pkt_handle_dc_eager_rtw_recv(struct efa_rdm_ep *ep,
-				      struct efa_rdm_pke *pkt_entry);
-
-void rxr_pkt_handle_longcts_rtw_recv(struct efa_rdm_ep *ep,
-				  struct efa_rdm_pke *pkt_entry);
-
-void rxr_pkt_handle_longread_rtw_recv(struct efa_rdm_ep *ep,
-				  struct efa_rdm_pke *pkt_entry);
 struct efa_rdm_ope *rxr_pkt_get_msgrtm_rxe(struct efa_rdm_ep *ep,
 						 struct efa_rdm_pke **pkt_entry_ptr);
